refactor: split core iteration kernel helpers
